Move location map into separate component
